feat: Apply new algorythm to get random quote
A new QuoteService file has been added to get a random quote from the
list of available ones. In order to avoid repeating quotes, this file
makes use of a number array that gets updated everytime a quote is
displayed. This array contains numbers that will represent the index to
each quote. When a quote is displayed, said index is extracted from the
array. The process repeats itself until there are no more numbers in the
list. By applying this logic, a quotes cannot be repeated unless all of the
others have been shown.

Once the list of numbers is empty, it will be populated again with
as many numbers as quotes available in the list. All of the previously
described logic is repeated all over, so all of the quotes must have
been seen for a second time before any of them can appear for a third
time and so on.